third_party/blink/web_tests/external/wpt/editing/run/fontsize_1-1000-expected.txt

This is a testharness.js-based test.
Found 20 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<p>[foo</p> <p>bar]</p>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-size:large\\">foo</span></p><p> <span style=\\"font-size:large\\"><span>bar</span></span> </p><p><span style=\\"font-size:large\\">baz</span></p>" but got "<p><span style=\\"font-size:large\\">foo</span></p><p> <span style=\\"font-size:large\\"><span>bar</span> </span></p><p><span style=\\"font-size:large\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"4\\">foo</font></p><p> <font size=\\"4\\"><span>bar</span></font> </p><p><font size=\\"4\\">baz</font></p>" but got "<p><font size=\\"4\\">foo</font></p><p> <font size=\\"4\\"><span>bar</span> </font></p><p><font size=\\"4\\">baz</font></p>"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<p>[foo<p><br><p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-size:large\\">foo</span></p><p><br></p><p><span style=\\"font-size:large\\">bar</span></p>" but got "<p><span style=\\"font-size:large\\">foo</span></p><p><span style=\\"font-size:large\\"><br></span></p><p><span style=\\"font-size:large\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<p>[foo<p><br><p>bar]" queryCommandIndeterm("fontsize") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<p>[foo<p><br><p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"4\\">foo</font></p><p><br></p><p><font size=\\"4\\">bar</font></p>" but got "<p><font size=\\"4\\">foo</font></p><p><font size=\\"4\\"><br></font></p><p><font size=\\"4\\">bar</font></p>"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<p>[foo<p><br><p>bar]" queryCommandIndeterm("fontsize") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontsize","7"]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"7\\">bar</font>baz" but got "foo<span style=\\"font-size:xxx-large\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["fontsize","8"]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"7\\">bar</font>baz" but got "foo<span style=\\"font-size:xxx-large\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["fontsize","100"]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"7\\">bar</font>baz" but got "foo<span style=\\"font-size:xxx-large\\">bar</span>baz"
[FAIL] [["fontsize","2em"]] "foo[bar]baz": execCommand("fontsize", false, "2em") return value
  assert_equals: expected false but got true
[FAIL] [["fontsize","2em"]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<font size=\\"2\\">bar</font>baz"
[FAIL] [["fontsize","2em"]] "foo[bar]baz" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "3" but got "2"
[FAIL] [["fontsize","20pt"]] "foo[bar]baz": execCommand("fontsize", false, "20pt") return value
  assert_equals: expected false but got true
[FAIL] [["fontsize","20pt"]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<font size=\\"7\\">bar</font>baz"
[FAIL] [["fontsize","20pt"]] "foo[bar]baz" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "3" but got "7"
[FAIL] [["fontsize","1."]] "foo[bar]baz": execCommand("fontsize", false, "1.") return value
  assert_equals: expected false but got true
[FAIL] [["fontsize","1."]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<font size=\\"1\\">bar</font>baz"
[FAIL] [["fontsize","1."]] "foo[bar]baz" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "3" but got "1"
[FAIL] [["stylewithcss","true"],["fontsize","+9"]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"7\\">bar</font>baz" but got "foo<span style=\\"font-size:xxx-large\\">bar</span>baz"
Harness: the test ran to completion.
